requirements.txt
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
elastic-transport==8.4.0
elasticsearch==8.9.0
exceptiongroup==1.1.3
flasgger==0.9.7.1
Flask==2.3.2
flask-sqlalchemy==3.0.5
greenlet==2.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
Mako==1.2.4
MarkupSafe==2.1.3
mistune==3.0.1
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.7
pytest==7.4.0
python-dotenv==1.0.0
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.9.2
six==1.16.0
SQLAlchemy==2.0.20
tomli==2.0.1
typing-extensions==4.7.1
urllib3==1.26.16
werkzeug==2.3.7
